requirements.txt
Flask
Flask-SocketIO
Flask-Cors
eventlet
pyserial
PyJWT